utilize useBoxLayout hook for horizontal separator logic
